Set up minimal test example for solver
The minimal example borrows from the tests for gala agglomeration.
The longer test is still available but only runs when setting
environment variable GALA_TEST_FULL. This should make the test suite
much much quicker.
